Module cadence::prelude

Export commonly used parts of Cadence for easy glob imports

Example

use cadence::prelude::*;
use cadence::{StatsdClient, NopMetricSink};

let client = StatsdClient::from_sink("some.prefix", NopMetricSink);

client.count("some.counter", 1).unwrap();
client.time("some.timer", 23).unwrap();
client.gauge("some.gauge", 45).unwrap();
client.meter("some.meter", 67).unwrap();
client.histogram("some.histogram", 89).unwrap();

Traits

Counted
Trait for incrementing and decrementing counters.

Gauged
Trait for recording gauge values.

Histogrammed
Trait for recording histogram values.

Metered
Trait for recording meter values.

MetricClient
Trait that encompasses all other traits for sending metrics.

Setted
Trait for recording set values.

Timed
Trait for recording timings in milliseconds.