Trait cadence::prelude::Gauged

pub trait Gauged {
    fn gauge_with_tags<'a>(
        &'a self, 
        key: &'a str, 
        value: u64
    ) -> MetricBuilder<Gauge>;

    fn gauge(&self, key: &str, value: u64) -> MetricResult<Gauge> { ... }
}

Trait for recording gauge values.

Gauge values are an instantaneous measurement of a value determined by the client. They do not change unless changed by the client. Examples include things like load average or how many connections are active.

See the Statsd spec for more information.

Note that tags are a Datadog extension to Statsd and may not be supported by your server.

Required Methods

fn gauge_with_tags<'a>(
&'a self,
key: &'a str,
value: u64
) -> MetricBuilder<Gauge>

Record a gauge value with the given key and return a MetricBuilder that can be used to add tags to the metric.

Provided Methods

fn gauge(&self, key: &str, value: u64) -> MetricResult<Gauge>

Record a gauge value with the given key