Struct cadence::BufferedUdpMetricSink

pub struct BufferedUdpMetricSink { /* fields omitted */ }

Implementation of a MetricSink that buffers metrics before sending them to a UDP socket.

Metrics are line buffered, meaning that a trailing "\n" is added after each metric written to this sink. When the buffer is sufficiently full, the contents of the buffer are flushed to a UDP socket and then the metric is written to the buffer. The buffer is also flushed when this sink is destroyed.

The default size of the buffer is 512 bytes. This is the "safest" size for a UDP packet according to the Etsy Statsd docs. The buffer size can be customized using the with_capacity method to create the sink if desired.

If a metric larger than the buffer is emitted, it will be written directly to the underlying UDP socket, bypassing the buffer.

Methods

impl BufferedUdpMetricSink

fn from<A>(sink_addr: A, socket: UdpSocket) -> MetricResult<BufferedUdpMetricSink> where A: ToSocketAddrs

Construct a new BufferedUdpMetricSink instance with a default buffer size of 512 bytes.

The address should be the address of the remote metric server to emit metrics to over UDP. The socket should already be bound to a local address with any desired configuration applied (blocking vs non-blocking, timeouts, etc.).

Writes to this sink are automatically suffixed with a Unix newline ('\n') by the sink and stored in a 512 byte buffer until the buffer is full or this sink is destroyed, at which point the buffer will be flushed.

Example

use std::net::UdpSocket;
use cadence::{BufferedUdpMetricSink, DEFAULT_PORT};

let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
let host = ("metrics.example.com", DEFAULT_PORT);
let sink = BufferedUdpMetricSink::from(host, socket);Run

Failures

This method may fail if:

It is unable to resolve the hostname of the metric server.

The host address is otherwise unable to be parsed

fn with_capacity<A>(sink_addr: A, socket: UdpSocket, cap: usize) -> MetricResult<BufferedUdpMetricSink> where A: ToSocketAddrs

Construct a new BufferedUdpMetricSink instance with a custom buffer size.

The address should be the address of the remote metric server to emit metrics to over UDP. The socket should already be bound to a local address with any desired configuration applied (blocking vs non-blocking, timeouts, etc.).

Writes to this sink are automatically suffixed with a Unix newline ('\n') by the sink and stored in a buffer until the buffer is full or this sink is destroyed, at which point the buffer will be flushed.

For guidance on sizing your buffer see the Statsd docs.

Example

use std::net::UdpSocket;
use cadence::{BufferedUdpMetricSink, DEFAULT_PORT};

let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
let host = ("metrics.example.com", DEFAULT_PORT);
let sink = BufferedUdpMetricSink::with_capacity(host, socket, 1432);Run

Failures

This method may fail if:

It is unable to resolve the hostname of the metric server.

The host address is otherwise unable to be parsed