Struct cadence::LoggingMetricSink

pub struct LoggingMetricSink {
    // some fields omitted
}

Implementation of a MetricSink that emits metrics using thelog! macro.

Metrics are emitted using the LogLevel provided at construction with a target of cadence::metrics. Note that the number of bytes written returned by emit does not reflect if the provided log level is high enough to be active.

Methods

impl LoggingMetricSink

fn new(level: LogLevel) -> LoggingMetricSink