Struct cadence::StatsdClientBuilder

pub struct StatsdClientBuilder { /* fields omitted */ }

Builder for creating and customizing StatsdClient instances.

Instances of the builder should be created by calling the ::builder() method on the StatsClient struct.

Example

use cadence::prelude::*;
use cadence::{MetricError, StatsdClient, NopMetricSink};

fn my_error_handler(err: MetricError) {
    println!("Metric error! {}", err);
}

let client = StatsdClient::builder("prefix", NopMetricSink)
    .with_error_handler(my_error_handler)
    .build();

client.count("something", 123);
client.count_with_tags("some.counter", 42)
    .with_tag("region", "us-east-2")
    .send();

Methods

impl StatsdClientBuilder

pub fn with_error_handler<F>(self, errors: F) -> Self where
F: Fn(MetricError) + Sync + Send + 'static,

Set an error handler to use for metrics sent via MetricBuilder::send()

The error handler is only invoked when metrics are not able to be sent correctly. Either due to invalid input, I/O errors encountered when trying to send them via a MetricSink, or some other reason.

The error handler should consume the error without panicking. The error may be logged, printed to stderr, discarded, etc. - this is up to the implementation.

pub fn build(self) -> StatsdClient

Construct a new StatsdClient instance based on current settings.