Trait cadence::Metric

pub trait Metric {
    fn as_metric_str(&self) -> &str;
}

Trait for metrics to expose Statsd metric string slice representation.

Implementing metrics know how to turn themselves into one of the supported types of metrics as defined in the Statsd spec.

Required methods

fn as_metric_str(&self) -> &str