Trait cadence::Timed

pub trait Timed {
    fn time(&self, key: &str, time: u64) -> MetricResult<Timer>;
}

Trait for recording timings in milliseconds.

Timings are a positive number of milliseconds between a start and end time. Examples include time taken to render a web page or time taken for a database call to return.

See the Statsd spec for more information.

Required Methods

fn time(&self, key: &str, time: u64) -> MetricResult<Timer>

Record a timing in milliseconds with the given key