Struct cadence::StatsdClient

pub struct StatsdClient<T: MetricSink> { /* fields omitted */ }

Client for Statsd that implements various traits to record metrics.

The client is the main entry point for users of this library. It supports several traits for recording metrics of different types.

Counted for emitting counters.

Timed for emitting timings.

Gauged for emitting gauge values.

Metered for emitting meter values.

MetricClient for a combination of all of the above.

For more information about the uses for each type of metric, see the documentation for each mentioned trait.

The client uses some implementation of a MetricSink to emit the metrics. In most cases, users will want to use the UdpMetricSink implementation.

Methods

impl<T: MetricSink> StatsdClient<T>

fn from_sink(prefix: &str, sink: T) -> StatsdClient<T>

Create a new client instance that will use the given prefix for all metrics emitted to the given MetricSink implementation.

No-op Example

use cadence::{StatsdClient, NopMetricSink};

let prefix = "my.stats";
let client = StatsdClient::from_sink(prefix, NopMetricSink);Run

UDP Socket Example

use std::net::UdpSocket;
use cadence::{StatsdClient, UdpMetricSink, DEFAULT_PORT};

let prefix = "my.stats";
let host = ("127.0.0.1", DEFAULT_PORT);

let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
socket.set_nonblocking(true).unwrap();

let sink = UdpMetricSink::from(host, socket).unwrap();
let client = StatsdClient::from_sink(prefix, sink);Run

Buffered UDP Socket Example

use std::net::UdpSocket;
use cadence::{StatsdClient, BufferedUdpMetricSink, DEFAULT_PORT};

let prefix = "my.stats";
let host = ("127.0.0.1", DEFAULT_PORT);

let socket = UdpSocket::bind("0.0.0.0:0").unwrap();

let sink = BufferedUdpMetricSink::from(host, socket).unwrap();
let client = StatsdClient::from_sink(prefix, sink);Run

fn from_udp_host<A>(prefix: &str, host: A) -> MetricResult<StatsdClient<UdpMetricSink>> where A: ToSocketAddrs

Create a new client instance that will use the given prefix to send metrics to the given host over UDP using an appropriate sink. This is the construction method that most users of this library will use.

The created UDP socket will be put into non-blocking mode.

Note that you must include a type parameter when you call this method to help the compiler determine the type of T (the sink).

Example

use cadence::{StatsdClient, UdpMetricSink};

let prefix = "my.stats";
let host = ("metrics.example.com", 8125);

// Note that we include a type parameter for the method call
let client = StatsdClient::<UdpMetricSink>::from_udp_host(prefix, host);Run

Failures

This method may fail if:

It is unable to create a local UDP socket.

It is unable to put the UDP socket into non-blocking mode.

It is unable to resolve the hostname of the metric server.

The host address is otherwise unable to be parsed.