Struct cadence::UdpMetricSink

pub struct UdpMetricSink {
    // some fields omitted
}

Implementation of a MetricSink that emits metrics over UDP.

This is the MetricSink that almost all consumers of this library will want to use. It accepts a UDP socket instance over which to write metrics and the address of the Statsd server to send packets to.

Methods

impl UdpMetricSink

fn from<A>(sink_addr: A, socket: UdpSocket) -> MetricResult<UdpMetricSink> where A: ToSocketAddrs

Construct a new UdpMetricSink instance.

The address should be the address of the remote metric server to emit metrics to over UDP. The socket should already be bound to a local address.

Example

use std::net::UdpSocket;
use cadence::{UdpMetricSink, DEFAULT_PORT};

let host = ("metrics.example.com", DEFAULT_PORT);
let socket = UdpSocket::bind("0.0.0.0:0").unwrap();
let sink = UdpMetricSink::from(host, socket);

Failures

This method may fail if:

It is unable to resolve the hostname of the metric server.

The host address is otherwise unable to be parsed